refactor(chat): trocear ChatToolCard en módulos tool-card/ — fase 1 (03/T02)
ChatToolCard.tsx baja de 1.298 a 790 líneas extrayendo sin cambios de
comportamiento: toolCardConfig.ts (mapeo tool→categoría/icono, límites
de codegen), toolResultParsers.ts (parsers puros de resultados y args) y
ToolResultHighlights.tsx (highlights de calendar/flashcard/resource,
preview de codegen, resumen de tree). API pública intacta: ToolCallData,
ChatToolSurfaceVariant, SubagentToolSection y ChatToolCardGroup siguen
exportándose desde ChatToolCard.tsx — los 10 consumidores no se tocan.
Siguientes (un componente por PR): FolderTabView, Runs/Automations
WorkspaceView, ManyPanel, UnifiedSidebar.
