+++ app/components/chat/ChatToolCard.tsx
@@ -2,48 +2,25 @@
 import { useState, useMemo, type ReactNode } from 'react';
 import { useTranslation } from 'react-i18next';
 import {
-  Globe,
-  Search,
   FileText,
   Loader2,
   CheckCircle2,
   XCircle,
   Check,
   ChevronRight,
-  Database,
-  Plug,
-  FileTextIcon,
-  Image,
   PlusCircle,
-  Calendar,
-  ShoppingBag,
-  GitBranch,
-  Crop,
-  Layers,
-  Bot,
-  Zap,
-  Network,
-  GraduationCap,
-  FileCode2,
-  FilePlus2,
-  FilePenLine,
-  Terminal,
-  FolderTree,
-  FolderSearch,
   Users,
 } from 'lucide-react';
 import MarkdownRenderer from './MarkdownRenderer';
-import ArtifactCard, { type AnyArtifact, type ArtifactType } from './ArtifactCard';
+import ArtifactCard from './ArtifactCard';
 import ChatTodoList, { parseTodos } from './ChatTodoList';
-import { tryParseArtifact, ZOD_VALIDATED_ARTIFACT_TYPES } from '@/lib/chat/artifactSchemas';
 import { useManyStore } from '@/lib/store/useManyStore';
 import { parseContentImages, parseImageResult } from '@/lib/chat/image-tool-utils';
 import DomeCollapsibleRow from '@/components/ui/DomeCollapsibleRow';
 import DomeButton from '@/components/ui/DomeButton';
 import DomeBadge from '@/components/ui/DomeBadge';
 import { getSubagentDisplayLabel } from '@/lib/chat/toolCatalog';
-import { getToolDisplayLabel, getToolDisplayLabelForCall } from '@/lib/chat/toolDisplayLabels';
-import { extractCalendarEventFromToolResult, unwrapToolResultPayload } from '@/lib/chat/calendarToolArtifact';
+import { getToolDisplayLabelForCall } from '@/lib/chat/toolDisplayLabels';
 import { JsonPrettyPrinterRoot } from '@/lib/chat/jsonPrettyPrinter';
 import { isFilesystemTreeTool, parseTreeToolSummary } from '@/lib/chat/treeToolSummary';
 import { stableStringHash } from '@/lib/utils/stableStringHash';
@@ -71,513 +48,27 @@ interface ChatToolCardProps {
   surfaceVariant?: ChatToolSurfaceVariant;
 }
 
-type ToolCategory = 'search' | 'file' | 'agent' | 'db' | 'mcp' | 'default';
+// Config, parsers y highlights extraídos (03/T02) — misma API pública.
+import {
+  type ToolCategory,
+  CATEGORY_COLORS,
+  getCategory,
+  getIconForTool,
+} from './tool-card/toolCardConfig';
+import {
+  parseDocumentResult,
+  parseArtifactResult,
+  parseResourceItems,
+  formatArgsSummary,
+  smartToolSummary,
+  getCodegenPreview,
+} from './tool-card/toolResultParsers';
+import {
+  renderToolSuccessHighlight,
+  CodegenPreview,
+  renderTreeToolSummary,
+} from './tool-card/ToolResultHighlights';
 
-/** Category color accent (border/dot color) using CSS variables for theme compatibility */
-const CATEGORY_COLORS: Record<ToolCategory, string> = {
-  search: 'var(--accent)',   // blue
-  file: 'var(--success)',     // green
-  agent: 'var(--accent)',    // purple
-  db: 'var(--warning)',       // orange
-  mcp: 'var(--secondary-text)',      // gray
-  default: 'var(--secondary-text)',  // gray
-};
-
-function getCategory(name: string): ToolCategory {
-  const n = (name || '').toLowerCase();
-  if (n.includes('search') || n.includes('web_fetch') || n.includes('web_search') || n.includes('fetch')) return 'search';
-  if (n.includes('calendar')) return 'db';
-  if (n.includes('marketplace')) return 'mcp';
-  if (n.includes('flashcard')) return 'file';
-  if (n.includes('pdf') || n.includes('file') || n.includes('resource') || n.includes('image') || n.includes('notebook')) return 'file';
-  if (n === 'glob' || n === 'ls' || n.includes('shell') || n.includes('codegen')) return 'file';
-  if (n === 'task' || n.includes('subagent') || n.includes('agent') || n.includes('call_') || n.includes('delegate')) return 'agent';
-  if (n.includes('postgres') || n.includes('sql') || n.includes('query') || n.includes('database') || n.includes('db')) return 'db';
-  if (n.startsWith('mcp') || n.includes('mcp_')) return 'mcp';
-  return 'default';
-}
-
-const TOOL_ICONS: Record<string, typeof Globe> = {
-  web_search: Search,
-  web_fetch: Globe,
-  deep_research: Search,
-  file_search: FolderSearch,
-  delegate_to_agent: Users,
-  resource_create: FileText,
-  resource_get: FileText,
-  resource_search: Search,
-  call_research_agent: Search,
-  call_library_agent: FileText,
-  call_writer_agent: FileText,
-  call_data_agent: Database,
-  start_async_subagent_task: GitBranch,
-  check_async_subagent_task: GitBranch,
-  update_async_subagent_task: GitBranch,
-  cancel_async_subagent_task: GitBranch,
-  list_async_subagent_tasks: GitBranch,
-  notebook_add_cell: FileText,
-  notebook_update_cell: FileText,
-  notebook_delete_cell: FileText,
-  pdf_extract_text: FileTextIcon,
-  pdf_get_metadata: FileTextIcon,
-  pdf_get_structure: FileTextIcon,
-  pdf_summarize: FileTextIcon,
-  pdf_extract_tables: FileTextIcon,
-  pdf_render_page: Image,
-  marketplace_search: ShoppingBag,
-  marketplace_install: ShoppingBag,
-  browser_get_active_tab: Globe,
-  workflow_create: GitBranch,
-  agent_create: Bot,
-  automation_create: Zap,
-  image_crop: Crop,
-  image_thumbnail: Layers,
-  generate_mindmap: Network,
-  generate_quiz: GraduationCap,
-  generate_knowledge_graph: Network,
-  calendar_list_events: Calendar,
-  calendar_list: Calendar,
-  calendar_get_upcoming: Calendar,
-  calendar_create: Calendar,
-  calendar_create_event: Calendar,
-  calendar_update: Calendar,
-  calendar_update_event: Calendar,
-  calendar_delete: Calendar,
-  calendar_delete_event: Calendar,
-  flashcard_create: Layers,
-  // Filesystem / codegen (deepagents harness + Dome aliases)
-  write_file: FilePlus2,
-  file_write: FilePlus2,
-  edit_file: FilePenLine,
-  read_file: FileCode2,
-  file_read: FileCode2,
-  glob: FolderSearch,
-  ls: FolderTree,
-  file_list: FolderTree,
-  file_tree: FolderTree,
-  shell_exec: Terminal,
-  // Subagent delegation (deepagents `task`)
-  task: Users,
-};
-
-function renderToolSuccessHighlight(
-  toolName: string,
-  rawResult: unknown,
-  t: (key: string, opts?: Record<string, unknown> & { defaultValue?: string }) => string,
-): ReactNode | null {
-  const cal = extractCalendarEventFromToolResult(toolName, rawResult);
-  if (cal) {
-    return (
-      <div
-        className="rounded-md border p-2.5 space-y-1"
-        style={{
-          borderColor: 'var(--border)',
-          background: 'color-mix(in srgb, var(--accent) 8%, transparent)',
-        }}
-      >
-        <div className="flex items-center gap-2 text-xs font-semibold" style={{ color: 'var(--primary-text)' }}>
-          <Calendar className="size-3.5 shrink-0 text-[var(--accent)]" aria-hidden />
-          <span className="truncate">{cal.title || t('chat.calendar_event_untitled', { defaultValue: 'Evento' })}</span>
-        </div>
-        {cal.startLabel ? (
-          <p className="text-[12px]" style={{ color: 'var(--secondary-text)' }}>
-            {cal.startLabel}
-            {cal.endLabel && cal.endLabel !== cal.startLabel ? ` → ${cal.endLabel}` : ''}
-          </p>
-        ) : null}
-        {cal.location ? (
-          <p className="text-[12px]" style={{ color: 'var(--tertiary-text)' }}>
-            {cal.location}
-          </p>
-        ) : null}
-        {cal.id ? (
-          <p className="text-[12px] font-mono opacity-70 truncate" style={{ color: 'var(--tertiary-text)' }}>
-            {cal.id}
-          </p>
-        ) : null}
-      </div>
-    );
-  }
-
-  const parsed = unwrapToolResultPayload(rawResult);
-  if (!parsed) return null;
-  const n = (toolName || '').toLowerCase();
-  const ok = parsed.success === true || parsed.status === 'success';
-
-  if (n === 'flashcard_create' && ok && parsed.deck && typeof parsed.deck === 'object') {
-    const deck = parsed.deck as Record<string, unknown>;
-    const title = String(deck.title || '');
-    const count = typeof deck.card_count === 'number' ? deck.card_count : 0;
-    return (
-      <div
-        className="rounded-md border p-2.5 space-y-1"
-        style={{
-          borderColor: 'var(--border)',
-          background: 'color-mix(in srgb, var(--success) 8%, transparent)',
-        }}
-      >
-        <div className="flex items-center gap-2 text-xs font-semibold" style={{ color: 'var(--primary-text)' }}>
-          <Layers className="size-3.5 shrink-0 text-[var(--success)]" aria-hidden />
-          <span className="truncate">{title}</span>
-        </div>
-        <p className="text-[12px]" style={{ color: 'var(--secondary-text)' }}>
-          {t('chat.flashcard_deck_count', { count, defaultValue: '{{count}} tarjetas' })}
-        </p>
-      </div>
-    );
-  }
-
-  if (n === 'resource_create' && ok && parsed.resource && typeof parsed.resource === 'object') {
-    const r = parsed.resource as Record<string, unknown>;
-    const title = String(r.title || '');
-    const id = String(r.id || '');
-    const typ = String(r.type || '');
-    return (
-      <div
-        className="rounded-md border p-2.5 flex gap-2 items-start"
-        style={{ borderColor: 'var(--border)', background: 'var(--bg-tertiary)' }}
-      >
-        <FileText className="size-3.5 shrink-0 mt-0.5 text-[var(--accent)]" aria-hidden />
-        <div className="min-w-0">
-          <p className="text-xs font-semibold truncate" style={{ color: 'var(--primary-text)' }}>
-            {title}
-          </p>
-          <p className="text-[12px] font-mono opacity-70 truncate" style={{ color: 'var(--tertiary-text)' }}>
-            {typ} · {id}
-          </p>
-        </div>
-      </div>
-    );
-  }
-
-  const st = parsed.status;
-  if (st === 'success') {
-    const thumb = typeof parsed.thumbnail === 'string' ? parsed.thumbnail : '';
-    const cropped = typeof parsed.croppedImage === 'string' ? parsed.croppedImage : '';
-    const src = cropped || thumb;
-    if (src.startsWith('data:')) {
-      return (
-        <img
-          src={src}
-          alt=""
-          className="max-w-[220px] max-h-[160px] object-contain rounded-md border border-[var(--border)]"
-        />
-      );
-    }
-  }
-
-  return null;
-}
-
-function getIconForTool(name: string): typeof Globe {
-  const norm = (name || '').toLowerCase();
-  if (TOOL_ICONS[norm]) return TOOL_ICONS[norm];
-  if (TOOL_ICONS[name]) return TOOL_ICONS[name];
-  if (norm.includes('postgres') || norm.includes('sql') || norm.includes('query') || norm.includes('database')) return Database;
-  if (norm.includes('mcp_') || norm.startsWith('mcp')) return Plug;
-  return Globe;
-}
-
-/** Parse result as document-style array [{ content, metadata }] */
-function parseDocumentResult(result: unknown): Array<{ content?: string; metadata?: Record<string, unknown> }> | null {
-  if (!result) return null;
-  let parsed: unknown;
-  if (typeof result === 'string') {
-    try { parsed = JSON.parse(result); } catch { return null; }
-  } else {
-    parsed = result;
-  }
-  if (!Array.isArray(parsed) || parsed.length === 0) return null;
-  const valid = parsed.every(
-    (item) =>
-      item &&
-      typeof item === 'object' &&
-      (typeof (item as { content?: unknown }).content === 'string' ||
-        typeof (item as { metadata?: unknown }).metadata === 'object')
-  );
-  return valid ? (parsed as Array<{ content?: string; metadata?: Record<string, unknown> }>) : null;
-}
-
-/** Parse result as artifact */
-function parseArtifactResult(result: unknown): AnyArtifact | null {
-  if (!result) return null;
-  let parsed: unknown;
-  if (typeof result === 'string') {
-    try { parsed = JSON.parse(result); } catch { return null; }
-  } else if (result && typeof result === 'object') {
-    parsed = result;
-  } else {
-    return null;
-  }
-  if (!parsed || typeof parsed !== 'object') return null;
-  const obj = parsed as Record<string, unknown>;
-  let artifact: AnyArtifact | undefined;
-  if (obj.artifact && typeof obj.artifact === 'object') artifact = obj.artifact as AnyArtifact;
-  if (!artifact && obj.content && Array.isArray(obj.content)) {
-    const textContent = obj.content[0]?.text;
-    if (typeof textContent === 'string') {
-      try {
-        const p = JSON.parse(textContent);
-        if (p.artifact) artifact = p.artifact as AnyArtifact;
-      } catch { /* Not JSON */ }
-    }
-  }
-  if (!artifact && obj.details && typeof obj.details === 'object') {
-    const details = obj.details as Record<string, unknown>;
-    if (details.artifact) artifact = details.artifact as AnyArtifact;
-  }
-  if (!artifact) return null;
-  const artifactType = (artifact as { type?: string }).type as ArtifactType | undefined;
-  if (!artifactType) return null;
-  const legacyTypes: ArtifactType[] = [
-    'pdf_summary',
-    'table',
-    'action_items',
-    'chart',
-    'code',
-    'list',
-    'created_entity',
-    'docling_images',
-  ];
-  if (ZOD_VALIDATED_ARTIFACT_TYPES.has(artifactType)) {
-    const validated = tryParseArtifact(artifactType, artifact);
-    if (!validated.ok) return null;
-    return validated.value as AnyArtifact;
-  }
-  if (!legacyTypes.includes(artifactType)) return null;
-  return artifact;
-}
-
-interface ResourceItem {
-  id: string;
-  title: string;
-  type: string;
-  snippet?: string;
-  similarity?: number;
-}
-
-function parseResourceItems(toolName: string, result: unknown): ResourceItem[] | null {
-  const n = (toolName || '').toLowerCase();
-  if (!n.includes('resource_list') && !n.includes('resource_search') && !n.includes('resource_semantic')) return null;
-  let parsed: unknown;
-  if (typeof result === 'string') {
-    try { parsed = JSON.parse(result); } catch { return null; }
-  } else {
-    parsed = result;
-  }
-  if (!parsed || typeof parsed !== 'object') return null;
-  const obj = parsed as Record<string, unknown>;
-  const arr = Array.isArray(obj.results) ? obj.results : Array.isArray(obj.resources) ? obj.resources : null;
-  if (!arr) return null;
-  return arr
-    .filter((item): item is Record<string, unknown> => item && typeof item === 'object')
-    .map((item) => ({
-      id: String(item.id || ''),
-      title: String(item.title || '(sin título)'),
-      type: String(item.type || 'resource'),
-      snippet: typeof item.snippet === 'string' ? item.snippet : undefined,
-      similarity: typeof item.similarity === 'number' ? item.similarity : undefined,
-    }))
-    .filter((item) => item.id);
-}
-
-/** Format args as a short single-line summary, truncated at ~60 chars */
-function formatArgsSummary(args: Record<string, unknown>): string {
-  const parts = Object.entries(args || {})
-    .filter(([, v]) => v !== undefined && v !== null && v !== '')
-    .map(([key, value]) => `${key}: ${typeof value === 'string' ? value : JSON.stringify(value)}`);
-  const joined = parts.join(', ');
-  if (joined.length > 60) return joined.slice(0, 60) + '…';
-  return joined;
-}
-
-/** Human-readable one-liner summary for the many panel card style */
-function smartToolSummary(name: string, args: Record<string, unknown>): string {
-  const n = name.toLowerCase();
-  if (n === 'file_write' || n === 'write_file' || n === 'edit_file' || n.includes('resource_create') || n.includes('notebook')) {
-    const fp = String(args.file_path ?? args.path ?? '');
-    if (fp) return fp.split('/').slice(-2).join('/');
-    const title = String(args.title ?? '');
-    return title.length > 64 ? title.slice(0, 61) + '…' : title;
-  }
-  if (n === 'file_read' || n === 'read_file') {
-    const fp = String(args.file_path ?? args.path ?? '');
-    return fp ? fp.split('/').slice(-1)[0]! : 'file';
-  }
-  if (n === 'glob') return String(args.pattern ?? args.glob ?? '').slice(0, 64);
-  if (n === 'ls' || n === 'file_list' || n === 'file_tree') {
-    return String(args.file_path ?? args.path ?? args.dir ?? '').slice(0, 64);
-  }
-  if (n === 'task' || n === 'delegate_to_agent') {
-    const sub = String(args.subagent_type ?? args.subagentType ?? args.agent ?? args.name ?? '');
-    const desc = String(args.prompt ?? args.task ?? args.description ?? '');
-    if (sub && desc) return `${sub}: ${desc}`.slice(0, 72);
-    return (sub || desc).slice(0, 72);
-  }
-  if (n === 'shell_exec' || n.includes('shell')) {
-    const cmd = String(args.command ?? '').trim();
-    return cmd.length > 72 ? cmd.slice(0, 69) + '…' : cmd;
-  }
-  if (n.includes('web_search') || n.includes('resource_search') || n.includes('memory')) {
-    return `"${String(args.query ?? args.q ?? '').slice(0, 60)}"`;
-  }
-  if (n.includes('web_fetch')) return String(args.url ?? '').slice(0, 72);
-  if (n.includes('resource_get')) {
-    return String(args.title ?? args.resourceId ?? args.id ?? '').slice(0, 64);
-  }
-  if (n.includes('calendar')) {
-    return String(args.title ?? args.summary ?? '').slice(0, 64);
-  }
-  return formatArgsSummary(args);
-}
-
-const EXT_LANG: Record<string, string> = {
-  html: 'HTML', htm: 'HTML', css: 'CSS', scss: 'SCSS',
-  js: 'JS', jsx: 'JSX', mjs: 'JS', cjs: 'JS',
-  ts: 'TS', tsx: 'TSX', py: 'Python', rb: 'Ruby', go: 'Go',
-  rs: 'Rust', java: 'Java', json: 'JSON', md: 'Markdown',
-  sh: 'Shell', bash: 'Shell', yml: 'YAML', yaml: 'YAML', sql: 'SQL', toml: 'TOML',
-};
-
-const CODEGEN_MAX_LINES = 48;
-const CODEGEN_MAX_CHARS = 2400;
-
-/** Extract a code preview from a filesystem/codegen tool's arguments, or null. */
-function getCodegenPreview(
-  name: string,
-  args: Record<string, unknown>,
-): { path: string; code: string; lang: string; truncated: boolean } | null {
-  const n = (name || '').toLowerCase();
-  if (n !== 'write_file' && n !== 'file_write' && n !== 'edit_file') return null;
-  const path = String(args.file_path ?? args.path ?? '');
-  let code = '';
-  if (typeof args.content === 'string') code = args.content;
-  else if (typeof args.new_string === 'string') code = args.new_string;
-  else if (typeof args.text === 'string') code = args.text;
-  if (!code.trim()) return null;
-
-  const ext = path.includes('.') ? path.split('.').pop()!.toLowerCase() : '';
-  const lang = EXT_LANG[ext] ?? '';
-
-  const lines = code.split('\n');
-  let truncated = false;
-  let preview = code;
-  if (lines.length > CODEGEN_MAX_LINES) {
-    preview = lines.slice(0, CODEGEN_MAX_LINES).join('\n');
-    truncated = true;
-  }
-  if (preview.length > CODEGEN_MAX_CHARS) {
-    preview = preview.slice(0, CODEGEN_MAX_CHARS);
-    truncated = true;
-  }
-  return { path, code: preview, lang, truncated };
-}
-
-function CodegenPreview({
-  preview,
-  t,
-}: {
-  preview: { path: string; code: string; lang: string; truncated: boolean };
-  t: (key: string, opts?: Record<string, unknown> & { defaultValue?: string }) => string;
-}) {
-  const fileName = preview.path ? preview.path.split('/').slice(-1)[0] : '';
-  return (
-    <div
-      className="rounded-md border overflow-hidden"
-      style={{ borderColor: 'var(--border)', background: 'var(--bg-tertiary)' }}
-    >
-      <div
-        className="flex items-center gap-2 px-2.5 py-1.5 border-b"
-        style={{ borderColor: 'var(--border)' }}
-      >
-        <FileCode2 className="size-3.5 shrink-0" style={{ color: 'var(--success)' }} aria-hidden />
-        <span
-          className="text-[11.5px] font-mono truncate flex-1"
-          style={{ color: 'var(--secondary-text)' }}
-          title={preview.path}
-        >
-          {fileName || preview.path}
-        </span>
-        {preview.lang ? (
-          <DomeBadge label={preview.lang} variant="soft" size="xs" color="var(--tertiary-text)" className="shrink-0" />
-        ) : null}
-      </div>
-      <pre
-        style={{
-          fontSize: 11.5,
-          lineHeight: 1.5,
-          fontFamily: "'JetBrains Mono', 'Fira Code', monospace",
-          whiteSpace: 'pre',
-          overflowX: 'auto',
-          overflowY: 'auto',
-          maxHeight: 280,
-          color: 'var(--primary-text)',
-          padding: '8px 10px',
-          margin: 0,
-        }}
-      >
-        {preview.code}
-      </pre>
-      {preview.truncated ? (
-        <div
-          className="px-2.5 py-1 text-[11px] border-t"
-          style={{ borderColor: 'var(--border)', color: 'var(--tertiary-text)' }}
-        >
-          {t('chat.codegen_truncated', { defaultValue: '… vista previa truncada' })}
-        </div>
-      ) : null}
-    </div>
-  );
-}
-
-function renderTreeToolSummary(summary: ReturnType<typeof parseTreeToolSummary>, t: (key: string, opts?: { defaultValue?: string }) => string) {
-  if (!summary) return null;
-  return (
-    <div
-      style={{
-        display: 'flex',
-        flexDirection: 'column',
-        gap: 6,
-        fontSize: 12,
-        color: 'var(--secondary-text)',
-        padding: '8px 10px',
-        background: 'var(--bg-tertiary)',
-        borderRadius: 4,
-      }}
-    >
-      {summary.path ? (
-        <div>
-          <span style={{ fontWeight: 600, color: 'var(--primary-text)' }}>{t('chat.tree_tool_path', { defaultValue: 'Ruta' })}: </span>
-          <span style={{ wordBreak: 'break-all' }}>{summary.path}</span>
-        </div>
-      ) : null}
-      {summary.shown != null ? (
-        <div>
-          <span style={{ fontWeight: 600, color: 'var(--primary-text)' }}>{t('chat.tree_tool_entries', { defaultValue: 'Entradas' })}: </span>
-          {summary.shown}
-          {summary.truncated ? ` (${t('chat.tree_tool_truncated', { defaultValue: 'truncado' })})` : ''}
-        </div>
-      ) : null}
-      {summary.max_depth != null ? (
-        <div>
-          <span style={{ fontWeight: 600, color: 'var(--primary-text)' }}>{t('chat.tree_tool_depth', { defaultValue: 'Profundidad' })}: </span>
-          {summary.max_depth}
-        </div>
-      ) : null}
-      {summary.node_count != null ? (
-        <div>
-          <span style={{ fontWeight: 600, color: 'var(--primary-text)' }}>{t('chat.tree_tool_nodes', { defaultValue: 'Nodos' })}: </span>
-          {summary.node_count}
-        </div>
-      ) : null}
-      <p style={{ margin: 0, opacity: 0.85 }}>
-        {t('chat.tree_tool_hint', {
-          defaultValue: 'Usa file_list o file_tree acotado en lugar de directory_tree en carpetas grandes.',
-        })}
-      </p>
-    </div>
-  );
-}
 
 export default function ChatToolCard({ toolCall, className = '', surfaceVariant = 'default' }: ChatToolCardProps) {
   const { t } = useTranslation();
